Don't rely on the type hint to convert inputs to lists
I thought the previous solution of comparing the string version of the
type was not flexible enough. Instead, I treat every input that we got
multiple values for as a list. If we got a single value input and it
should've been considered a list with one element, handle that case
during generic parsing.